+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v3.8.0
 - Update minimum PHP version to 7.4
 - DataProviderStaticFixer - add option "force"
+- Deprecate InternalClassCasingFixer
 
 ## v3.7.0
 - Add NoTrailingCommaInSinglelineFixer